src/tools/pyproject.toml
[project]
name = "redpepper-tools"
description = "A state-based server management system (related tools package)"
readme = "README.md"
license = { text = "MIT License" }
requires-python = ">=3.12"
dynamic = ["version"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Information Technology",
  "License :: OSI Approved :: MIT License",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: Python :: 3",
  "Topic :: System :: Installation/Setup",
  "Topic :: System :: Systems Administration",
]
dependencies = [
    "argon2-cffi>=23.1.0",
    "pyotp>=2.9.0",
    "qrcode>=8.0",
    "requests>=2.32.3",
    "typer>=0.12.5",
]

[project.scripts]
redpepper-tools = "redpepper.tools.entrypoint:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "../redpepper/version.py"

[tool.hatch.build.targets.wheel]
packages = ["redpepper"]

[tool.ruff]
extend = "../../pyproject.toml"
